#24 - Found the issue
Now to actually fix it.

Did some refactorings while I was in here though.  Removed all instances of the word Jita and moved it to a constant for easy finding later.

+++ aba-industry/industry-common/src/test/java/com/aba/industry/invention/impl/InventionCalculatorImplUnitTests.java
@@ -61,8 +61,7 @@ public class InventionCalculatorImplUnitTests {
 
         for ( ActivityMaterialWithCost am : bpData.getActivityMaterials()
                                                   .get( IndustryActivities.INVENTION.getActivityId() ) ) {
-            ItemCost ic = itemCosts.get( am.getTypeId()
-                                           .intValue() );
+            ItemCost ic = itemCosts.get( am.getTypeId() );
 
             am.setCost( ic.getSell() );
             am.setAdjustedCost( ic.getAdjusted() );
@@ -70,8 +69,7 @@ public class InventionCalculatorImplUnitTests {
 
         for ( ActivityMaterialWithCost am : bpData.getActivityMaterials()
                                                   .get( IndustryActivities.MANUFACTURING.getActivityId() ) ) {
-            ItemCost ic = itemCosts.get( am.getTypeId()
-                                           .intValue() );
+            ItemCost ic = itemCosts.get( am.getTypeId() );
 
             am.setCost( ic.getSell() );
             am.setAdjustedCost( ic.getAdjusted() );
